Restructured dirs and updated plotting notebook
