+++ misc/irr-entities.txt
@@ -1,0 +1,299 @@
+Charles Schwab Corporation
+Nestle
+China's Tianqi Lithium Corporation
+LG Energy Solution Ltd
+Albemarle
+SQM
+Todd
+Boehringer Ingelheim
+Sagility
+Aboitiz Power
+Globe Group
+Procter & Gamble
+Unilever
+Philip Morris International
+Prudential
+Citi
+Kyndryl
+Mynt
+Accenture
+Manolo Lopez
+Forbes
+CIMB Bank
+Philippine National Bank (PNB)
+Bank of the Philippine Islands (BPI)
+Union Bank of the Philippines
+BDO Unibank
+Land Bank of the Philippines
+Philippine Savings Bank (PSBank)
+Rizal Commercial Banking Corporation (RCBC)
+Maya Bank
+Robinsons Bank
+Tan-owned PNB
+Ayala-led BPI
+Sy-owned BDO
+SNB
+Ryanair
+Airbus
+Boeing
+Netflix
+Ahlers
+Kering
+Gucci
+Codelco
+Summit Nanotech Corporation
+World Bank
+Yuchengco Group
+Rosimushchestvo
+Gazprom
+Uniper
+Fortum
+Argentina's economic minister
+American's Isom
+Delta Air Lines
+Rappler.com
+Japanese yen
+First Republic results
+European Central Bank policymakers
+Santander
+Alphabet
+Microsoft
+McDonald's
+United Parcel Service
+PepsiCo
+General Motors
+3M
+Silvergate Bank
+Casinos Austria International
+Sumitomo Mitsui Banking Corporation
+MUFG Bank
+Orix Corporation
+MGM Resorts International
+Ueda
+Bank of England
+UBS Group
+Federal Reserve's Board of Governors
+National Power Corporation
+Government Service Insurance System
+Philippine Ports Authority
+Commission on Audit
+Commission on Audit (COA)
+China Telecom
+Faurecia
+Embracer
+Nvidia
+Marvell Technology
+Bangko Sentral ng Pilipinas
+Ayala Land
+SVB
+China Huarong Asset Management Company Ltd
+NONE.
+Michelin
+KOREIN TILLERY
+PALLAS PARTNERS
+QUINN EMANUEL URQUHART & SULLIVAN
+Dcoop
+Luzales
+Volvo
+Yale
+Harvard
+Ruth
+Cargill
+Regions
+Huntington
+First Republic
+Anna
+Citizens Financial Group
+Stellantis
+British Steel
+S3 Partners
+Standard Chartered
+First Abu Dhabi Bank (FAB)
+Public Investment Fund
+Qatar Investment Authority (QIA)
+Swiss People's Party
+Credit Suisse leadership
+Deloitte
+BlackRock
+Deutsche Bank
+Swiss government
+Alantra
+Kone
+Grifols
+BASF
+Vodafone
+Telecom Italia
+Philips
+Logitech
+Zalando
+Sainsbury's
+ICA Gruppen
+Deliveroo
+Clas Ohlson
+Ahold Delhaize
+Lily
+Kumar
+LVMH
+Associated British Foods
+Just Eat Takeaway.com
+Burberry
+Cartier
+IG Metall
+Ford Motor Company
+Merck
+analysts at Industrial Securities
+German
+Beijing
+Russia
+OPEC+
+International Energy Agency
+deutsche telekom
+Daimler
+European Commission
+Volkswagen
+Audi
+BMW
+Mercedes-Benz
+Mitsubishi UFJ Financial Group (MUFG)
+Mizuho
+IPM Holdings
+MJC Investments
+TKC Metals
+Manila Jockey Club
+DFNN
+George Floyd's brother
+Pratt & Whitney
+trader at an international trading company
+Santa Clara
+Bangko Sentral ng Pilipinas (BSP)
+National Economic and Development Authority (NEDA)
+Philippine Statistics Authority
+Department of Environment and Natural Resources
+O-A
+Treasury Department-Head Office
+Jane Doe 1
+UBS
+Silicon Valley Bank
+Federal Deposit Insurance Corporation (FDIC)
+Silicon Valley Bank (SVB)
+JPMorgan Chase & Co.
+United Airlines
+SAP
+SVB Financial Group
+Maya
+Toyota Motor Corporation
+Bank of America
+Citigroup
+Reuters
+Washington Mutual
+ABS-CBN
+International Flavors & Fragrances (IFF)
+DSM
+Firmenich
+Symrise
+COA
+Primewater
+FINMA
+Sky
+Abu Dhabi state-backed investment vehicle ADQ
+Buzzfeed
+Sunlight Financial Holding
+AcuityAds Holdings
+Astra Space
+Cohu
+Rocket Lab USA
+Alkami Technology
+Stitch Fix
+Circle
+Rose
+Roku
+Monarchy
+Roblox
+European Central Bank
+Frankfurt, Germany
+Apollo Global Management
+Blackstone
+KKR & Co.
+Ares Management
+Carlyle Group
+BlockFi
+Payoneer Global
+Kinnevik AB
+Alecta
+Nitro Software Australia
+Vir Biotechnology
+Sunrun
+Sunnova Energy International
+Oak Street Health
+Ginkgo Bioworks Holdings
+Alphatec Holdings
+Trustpilot Group
+Diaceutics
+ECB policymakers
+Aldi
+Megaworld's
+Suntrust
+Tigre de Cristal
+Summit Ascent Investments Limited
+Rizal Commercial Banking Corporation (RCBC)
+Bangladesh Bank
+Ms. Zhu
+Zhang
+Wei
+Sun
+Duterte
+Ma
+Du's parents
+Li
+Zhou
+Aboitiz
+Marcos
+Powell
+Romeo
+Ye
+Diokno
+Zulai
+Yuriy
+TD Cowen
+Chris Burnet Ramos
+Joann Manabat
+Isagani de Castro Jr.
+Sofia Tomacruz
+Eirenne Lumasang
+Martha Teodoro
+Dennis Abrina
+Lance Spencer Yu
+Jodesz Gavilan
+Jairo Bolledo
+Ralf Rivas
+James Patrick Cruz
+Bea Cupin
+Lian Buan
+Lorenz Pasion
+Ryan Macasero
+Michelle Abad
+Bonz Magsambol
+Iya Gozum
+Lance Yu
+Chris Burnet Ramos/Rappler.com
+Joann Manabat/Rappler.com
+Isagani de Castro Jr./Rappler.com
+Sofia Tomacruz/Rappler.com
+Eirenne Lumasang/Rappler.com
+Martha Teodoro/Rappler.com
+Dennis Abrina/Rappler.com
+Lance Spencer Yu/Rappler.com
+Jodesz Gavilan/Rappler.com
+Jairo Bolledo/Rappler.com
+Ralf Rivas/Rappler.com
+James Patrick Cruz/Rappler.com
+Bea Cupin/Rappler.com
+Lian Buan/Rappler.com
+Lorenz Pasion/Rappler.com
+Ryan Macasero/Rappler.com
+Michelle Abad/Rappler.com
+Bonz Magsambol/Rappler.com
+Iya Gozum/Rappler.com
+Lance Yu/Rappler.com
+Lance Yu-Rappler.com
+Baron Antenna Digital
