+++ misc/labels.txt
@@ -1,0 +1,13 @@
+President
+Judge
+Dr
+Labour shadow finance minister
+OakNorth CEO
+Michelin-starred chef
+Police Colonel
+Governor
+Brother
+TD Cowen analyst
+NEDA Secretary
+CFRA Research
+Senator
